backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
pymysql==1.1.0
cryptography==41.0.7

# OpenAI and LangChain
openai==1.3.7
langchain==0.1.0
langchain-openai==0.0.2
tiktoken==0.5.2

# Streamlit
streamlit==1.28.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# Logging
python-json-logger==2.0.7
